+++ src/otel.ts
@@ -1,7 +1,8 @@
 import { ColumnHint, TimeUnit } from "types/queryBuilder";
 
 const defaultLogsTable = 'otel_logs';
-const defaultTraceTable = 'otel_traces'
+const defaultTraceTable = 'otel_traces';
+const defaultTraceTimestampTable = 'otel_traces_trace_id_ts';
 
 export interface OtelVersion {
   name: string;
@@ -11,6 +12,7 @@ export interface OtelVersion {
   logColumnMap: Map<ColumnHint, string>;
   logLevels: string[];
   traceTable: string;
+  traceTimestampTable: string;
   traceColumnMap: Map<ColumnHint, string>;
   traceDurationUnit: TimeUnit.Nanoseconds;
 }
@@ -35,6 +37,7 @@ const otel129: OtelVersion = {
     'FATAL'
   ],
   traceTable: defaultTraceTable,
+  traceTimestampTable: defaultTraceTimestampTable,
   traceColumnMap: new Map<ColumnHint, string>([
     [ColumnHint.Time, 'Timestamp'],
     [ColumnHint.TraceId, 'TraceId'],
@@ -56,3 +59,10 @@ export const versions: readonly OtelVersion[] = [
 ];
 
 export const getLatestVersion = (): OtelVersion => versions[0];
+export const getVersion = (version: string | undefined): OtelVersion | undefined => {
+  if (!version) {
+    return;
+  }
+
+  return versions.find(v => v.version === version);
+};